IntelliJ automated code cleanup
* Replace explicit type arguments with <>
* Add missing <p> tags to Javadoc
* Remove redundant warning suppressions
* Replace anonymous class with lambda
* Use try-with-resources
* Collapse identical catch blocks
* Use Comparator.comparing
* Use Map.getOrDefault
* Replace lambda invocation with body
* Other miscellaneous

+++ bson/src/main/org/bson/json/JsonStringBuffer.java
@@ -37,7 +37,7 @@ class JsonStringBuffer implements JsonBuffer {
             eof = true;
             return -1;
         } else {
-            return (int) buffer.charAt(position++);
+            return buffer.charAt(position++);
         }
     }
